fix: rename mock products file
